I [[borning]] added the patch from your code [[regarding picking a pivotable variable to pivot into the basis, instead of just any variable --gjb]], and also fixed another bug.  The new files are all on /homes/june/borning/smalltalk/* ... can you put these into the repository?
The other bug was in connection with symbolic weights, so it won't be in
the C++ or Java versions.  The bug was in the pivot method, in which it
found the inverse of a coefficient.  This didn't work right before if the
coefficient was a symbolic weight (1.0/w).  The only time this gets
tickled, as far as I can see, is if you delete all the constraints from a
variable (including any stays) so that it becomes completely
unconstrained.  Not likely to arise in practice, but it was cheap to fix.