Use file randoms.txt generatd by Alan to make benchmarks more comparable between languages.
Be sure to add exactly the number of cns requestd, not counting cns that could not be added;  these still are not perfect because the attempted adds take a lot of time (surprisingly much, it seems!)

Do the timings better, using timer.Reset() and timer.Stop().

Display the description of any caught exception.